bootloader: move Apple Event handling code to a separate file
