+++ news/6089.bootloader.rst
@@ -1,0 +1,1 @@
+Refactor Apple Events handling code and move it into a separate source file.
